requirements.txt
git+https://github.com/SmallsLIVE/smallslive.git@76892424aed97731c0925d075283803d6e830452
django-bmemcached==0.2.1
django-loaderio==0.3.0
django-oscar==0.7.2
django-oscar-paypal==0.9.1
django-postgrespool==0.3.0
django-sslify>=0.2
djangorestframework==2.4.8
gevent==1.0.1
gunicorn==19.1.1
natsort==3.5.0
newrelic==2.34.0.29
python-binary-memcached==0.24.2
raven==5.3.1
